R3.1 : Songs may be added by just a title. Offer a user-friendly way for the user to disambiguate options and choose the specific song they want (eg, user asking to add the song "Love").
R3.2 : When offering alternatives for disambiguation in R3.1, rank the suggestions intelligently (eg, based on popularity or similarity to the existing playlist).

R3.3 : Provide a possibility to ask questions about tracks and artists (support 4 types of question for either track or artist).

R3.4 : Provide a possibility to ask questions about compilation albums (eg Which artist appears most often in 'Best of 90s' albums)?

R3.5 : Provide textual or visual summary of the playlist with its basic statistics.